+++ app-sample/src/main/java/io/noties/markwon/app/sample/Tags.kt
@@ -2,4 +2,12 @@ package io.noties.markwon.app.sample
 
 object Tags {
     const val basics = "basics"
+    const val toast = "toast"
+    const val hack = "hack"
+    const val parsing = "parsing"
+    const val block = "block"
+    const val movementMethod = "movement-method"
+    const val links = "links"
+    const val plugin = "plugin"
+    const val recyclerView = "recycler-view"
 }